More wip for ParamConverter.
