+++ src/Criticalmass/DataQuery/Factory/ParamConverterFactoryInterface.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace App\Criticalmass\DataQuery\Factory;
+
+use Sensio\Bundle\FrameworkExtraBundle\Request\ParamConverter\ParamConverterInterface;
+
+interface ParamConverterFactoryInterface
+{
+    public function createParamConverter(string $fqcn): ?ParamConverterInterface;
+}
