+++ src/Criticalmass/DataQuery/Factory/ValueAssigner.php
@@ -6,18 +6,16 @@ use App\Criticalmass\DataQuery\Query\QueryInterface;
 use App\Criticalmass\DataQuery\QueryProperty\QueryProperty;
 use App\Criticalmass\Util\ClassUtil;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
-use Sensio\Bundle\FrameworkExtraBundle\Request\ParamConverter\ParamConverterInterface;
-use Symfony\Bridge\Doctrine\RegistryInterface;
 use Symfony\Component\HttpFoundation\Request;
 
 class ValueAssigner implements ValueAssignerInterface
 {
-    /** @var RegistryInterface $registry */
-    protected $registry;
+    /** @var ParamConverterFactoryInterface $paramConverterFactory */
+    protected $paramConverterFactory;
 
-    public function __construct(RegistryInterface $registry)
+    public function __construct(ParamConverterFactoryInterface $paramConverterFactory)
     {
-        $this->registry = $registry;
+        $this->paramConverterFactory = $paramConverterFactory;
     }
 
     public function assignPropertyValue(Request $request, QueryInterface $query, QueryProperty $property): QueryInterface
@@ -40,37 +38,25 @@ class ValueAssigner implements ValueAssignerInterface
                 break;
 
             default:
-                $query = $this->assignEntityValue($request, $query, $property);
+                $query = $this->assignEntityValueFromParamConverter($request, $query, $property);
                 break;
         }
 
         return $query;
     }
 
-    protected function assignEntityValue(Request $request, QueryInterface $query, QueryProperty $property): QueryInterface
+    protected function assignEntityValueFromParamConverter(Request $request, QueryInterface $query, QueryProperty $property): QueryInterface
     {
-        if ($converter = $this->createParamConverter($property->getType())) {
+        if ($converter = $this->paramConverterFactory->createParamConverter($property->getType())) {
             $methodName = $property->getMethodName();
+            $newParameterName = ClassUtil::getLowercaseShortnameFromFqcn($property->getType());
 
-            $paramConverterConfiguration = new ParamConverter(['name' => 'city']);
+            $paramConverterConfiguration = new ParamConverter(['name' => $newParameterName]);
 
             $converter->apply($request, $paramConverterConfiguration);
-            $query->$methodName($request->get('city'));
+            $query->$methodName($request->get($newParameterName));
         }
 
         return $query;
-    }
-
-    protected function createParamConverter(string $fqcn): ?ParamConverterInterface
-    {
-        $entityShortname = ClassUtil::getShortnameFromFqcn($fqcn);
-        $paramConverterNamespace = 'App\\Request\\ParamConverter\\';
-        $paramConverterFqcn = sprintf('%s%sParamConverter', $paramConverterNamespace, $entityShortname);
-
-        if (!class_exists($paramConverterFqcn)) {
-            return null;
-        }
-
-        return new $paramConverterFqcn($this->registry);
     }
 }
