+++ src/Criticalmass/DataQuery/QueryProperty/QueryProperty.php
@@ -63,6 +63,4 @@ class QueryProperty
 
         return $this;
     }
-
-
 }
